synthesis: route mc_* macros through runtime cells

+++ src/main/java/MinecraftGraph/MacroVertex.java
@@ -8,14 +8,14 @@ public class MacroVertex extends Function {
     private final String macroName;
     private final String outputPort;
     private final int outputBitIndex;
-    private final HashMap<String, Integer> params;
+    private final HashMap<String, Long> params;
 
-    public MacroVertex(int id, FunctionType type, int numInputs, String macroName, String outputPort, int outputBitIndex, Map<String, Integer> params) {
+    public MacroVertex(int id, FunctionType type, int numInputs, String macroName, String outputPort, int outputBitIndex, Map<String, Long> params) {
         super(id, type, numInputs);
         this.macroName = macroName;
         this.outputPort = outputPort;
         this.outputBitIndex = outputBitIndex;
-        this.params = new HashMap<String, Integer>(params);
+        this.params = new HashMap<String, Long>(params);
     }
 
     public String getMacroName() {
@@ -30,7 +30,7 @@ public class MacroVertex extends Function {
         return this.outputBitIndex;
     }
 
-    public Map<String, Integer> getParams() {
+    public Map<String, Long> getParams() {
         return Collections.unmodifiableMap(this.params);
     }
 }
